Changed so money values are done in pence using integers. Avoids doubles losing precision along the bytes...

+++ BankSimulator/src/com/main/account/Transaction.java
@@ -20,7 +20,7 @@ public class Transaction {
 	/*
 	 * Constructor when withdrawing money
 	 */
-	public Transaction(Choices c, double aux, int ac){
+	public Transaction(Choices c, int aux, int ac){
 		this.choice = c;
 		this.primaryAux = aux;
 		this.secondaryAux = ac;
